requirements.txt
fastapi[standard]
pydantic-settings
python-dotenv
sqlmodel
uvicorn